requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
jinja2>=3.1.6
requests>=2.32.0
Pillow>=11.1.0
pi-heif>=1.3.0
torchvision>=0.18.0
ultralytics>=8.3.0
python-dotenv>=1.0.1
